Rename Activity to ActivityLog Use wpcom v2 endpoint instead of v1 Move ActivityStore Errors, Payloads and Actions to inner classes
