+++ example/src/androidTest/java/org/wordpress/android/fluxc/release/ReleaseStack_AppComponent.java
@@ -45,5 +45,5 @@ public interface ReleaseStack_AppComponent {
     void inject(ReleaseStack_ThemeTestJetpack test);
     void inject(ReleaseStack_ThemeTestWPCom test);
     void inject(ReleaseStack_UploadTest test);
-    void inject(ReleaseStack_ActivityTestWPCom test);
+    void inject(ReleaseStack_ActivityLogTestWPCom test);
 }
