+++ fluxc/src/main/java/org/wordpress/android/fluxc/module/ReleaseNetworkModule.java
@@ -16,7 +16,7 @@ import org.wordpress.android.fluxc.network.discovery.DiscoveryWPAPIRestClient;
 import org.wordpress.android.fluxc.network.discovery.DiscoveryXMLRPCClient;
 import org.wordpress.android.fluxc.network.discovery.SelfHostedEndpointFinder;
 import org.wordpress.android.fluxc.network.rest.wpcom.account.AccountRestClient;
-import org.wordpress.android.fluxc.network.rest.wpcom.activity.ActivityRestClient;
+import org.wordpress.android.fluxc.network.rest.wpcom.activity.ActivityLogRestClient;
 import org.wordpress.android.fluxc.network.rest.wpcom.auth.AccessToken;
 import org.wordpress.android.fluxc.network.rest.wpcom.auth.AppSecrets;
 import org.wordpress.android.fluxc.network.rest.wpcom.auth.Authenticator;
@@ -155,10 +155,10 @@ public class ReleaseNetworkModule {
 
     @Singleton
     @Provides
-    public ActivityRestClient provideActivityRestClient(Context appContext, Dispatcher dispatcher,
-                                                        @Named("regular") RequestQueue requestQueue,
-                                                        AccessToken token, UserAgent userAgent) {
-        return new ActivityRestClient(appContext, dispatcher, requestQueue, token, userAgent);
+    public ActivityLogRestClient provideActivityRestClient(Context appContext, Dispatcher dispatcher,
+                                                           @Named("regular") RequestQueue requestQueue,
+                                                           AccessToken token, UserAgent userAgent) {
+        return new ActivityLogRestClient(appContext, dispatcher, requestQueue, token, userAgent);
     }
 
     @Singleton
