+++ fluxc/src/main/tools/wp-com-v2-endpoints.txt
@@ -1,2 +1,7 @@
 /users/username/suggestions/
 /plugins/featured
+
+/sites/$site/activity
+/sites/$site/rewind
+
+/activity-log/$site/rewind/to/$rewind_ID#String
